# app-judiciales/backend/docs/swagger.yaml
openapi: 3.0.3
info:
  title: Sistema de Expedientes Militares API
  description: |
    API REST para el sistema de gestión de expedientes militares
    
    ## Autenticación
    
    Esta API utiliza autenticación Bearer Token (JWT). Para usar los endpoints protegidos:
    
    1. **Primero, obtén un token** haciendo login en `/auth/login`
    2. **Autorízate en Swagger UI** haciendo clic en el botón "Authorize" 🔒 arriba
    3. **Ingresa el token** en el formato: `Bearer tu_token_aquí`
    4. **Haz clic en "Authorize"** y luego "Close"
    
    ### Ejemplo de autorización:
    ```
    Bearer eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9...
    ```
    
    ### Credenciales por defecto:
    - **Email:** admin@sistema.mil
    - **Password:** admin123
    
    Una vez autorizado, podrás usar todos los endpoints protegidos.
  version: 1.0.0
  contact:
    name: Soporte API
    email: soporte@sistema.mil
servers:
  - url: http://localhost:8082/api/v1
    description: Servidor de desarrollo
...
paths:
...
  /auth/login:
    post:
      tags:
        - Auth
      summary: Iniciar sesión
      description: |
        Autentica un usuario y devuelve un token JWT.
        
        **Pasos para usar la API:**
        1. Usa este endpoint para obtener tu token
        2. Copia el token de la respuesta
        3. Haz clic en "Authorize" 🔒 arriba de esta página
        4. Pega el token en el campo (formato: `Bearer tu_token`)
        5. Ya puedes usar todos los endpoints protegidos
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/LoginRequest'
            example:
              email: admin@sistema.mil
              password: admin123
      responses:
        '200':
          description: Login exitoso - Copia el token de la respuesta para autorización
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/LoginResponse'
        '401':
          description: Credenciales incorrectas
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /users:
    get:
      tags:
        - Users
      summary: Listar usuarios
      security:
        - BearerAuth: []
      responses:
        '200':
          description: Lista de usuarios
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                    example: true
                  data:
                    type: array
                    items:
                      $ref: '#/components/schemas/User'
        '401':
          description: No autorizado
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'

    post:
      tags:
        - Users
      summary: Crear usuario
      security:
        - BearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - email
                - nombre
                - apellido
                - documento
                - password
                - profile_id
              properties:
                email:
                  type: string
                  format: email
                  example: usuario@sistema.mil
                nombre:
                  type: string
                  example: Juan
                apellido:
                  type: string
                  example: Pérez
                documento:
                  type: string
                  example: 12345678
                password:
                  type: string
                  minLength: 8
                  example: password123
                profile_id:
                  type: string
                  example: 507f1f77bcf86cd799439011
      responses:
        '201':
          description: Usuario creado exitosamente
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                    example: true
                  data:
                    $ref: '#/components/schemas/User'
        '400':
          description: Error de validación
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '401':
          description: No autorizado
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /users/{id}:
    get:
      tags:
        - Users
      summary: Obtener usuario por ID
      security:
        - BearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Usuario encontrado
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                    example: true
                  data:
                    $ref: '#/components/schemas/User'
        '404':
          description: Usuario no encontrado
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '401':
          description: No autorizado
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'

    put:
      tags:
        - Users
      summary: Actualizar usuario
      security:
        - BearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                email:
                  type: string
                  format: email
                nombre:
                  type: string
                apellido:
                  type: string
                documento:
                  type: string
                activo:
                  type: boolean
      responses:
        '200':
          description: Usuario actualizado exitosamente
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                    example: true
                  data:
                    $ref: '#/components/schemas/User'
        '400':
          description: Error de validación
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '404':
          description: Usuario no encontrado
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'

    delete:
      tags:
        - Users
      summary: Eliminar usuario
      security:
        - BearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Usuario eliminado exitosamente
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SuccessResponse'
        '404':
          description: Usuario no encontrado
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /users/profile:
    get:
      tags:
        - Users
      summary: Obtener perfil del usuario actual
      description: Obtiene la información del perfil del usuario autenticado
      security:
        - BearerAuth: []
      responses:
        '200':
          description: Perfil del usuario actual
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                    example: true
                  data:
                    $ref: '#/components/schemas/User'
        '401':
          description: No autorizado
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'

    put:
      tags:
        - Users
      summary: Actualizar perfil del usuario actual
      description: Permite al usuario actualizar su propia información de perfil
      security:
        - BearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                nombre:
                  type: string
                  example: "Juan Carlos"
                apellido:
                  type: string
                  example: "Pérez García"
                telefono:
                  type: string
                  example: "+591 70123456"
              example:
                nombre: "Juan Carlos"
                apellido: "Pérez García"  
                telefono: "+591 70123456"
      responses:
        '200':
          description: Perfil actualizado exitosamente
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                    example: true
                  data:
                    $ref: '#/components/schemas/User'
                  message:
                    type: string
                    example: "perfil actualizado exitosamente"
        '400':
          description: Error de validación
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '401':
          description: No autorizado
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /users/password:
    put:
      tags:
        - Users
      summary: Cambiar contraseña del usuario actual
      description: Permite al usuario cambiar su propia contraseña
      security:
        - BearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - current_password
                - new_password
              properties:
                current_password:
                  type: string
                  example: "password123"
                  description: "Contraseña actual del usuario"
                new_password:
                  type: string
                  minLength: 8
                  example: "newPassword456"
                  description: "Nueva contraseña (mínimo 8 caracteres)"
              example:
                current_password: "password123"
                new_password: "newPassword456"
      responses:
        '200':
          description: Contraseña actualizada exitosamente
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                    example: true
                  message:
                    type: string
                    example: "contraseña actualizada exitosamente"
        '400':
          description: Error de validación o contraseña actual incorrecta
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '401':
          description: No autorizado
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /expedientes:
    get:
      tags:
        - Expedientes
      summary: Listar expedientes militares
      security:
        - BearerAuth: []
      parameters:
        - name: page
          in: query
          schema:
            type: integer
            default: 1
        - name: limit
          in: query
          schema:
            type: integer
            default: 10
        - name: grado
          in: query
          schema:
            type: string
            enum: [GRAL, CRL, TTE CRL, MY, CAP, TTE, STTE, TCO, SSOO, EC, TROPA]
        - name: situacion_militar
          in: query
          schema:
            type: string
            enum: [Actividad, Retiro]
        - name: estado
          in: query
          schema:
            type: string
            enum: [dentro, fuera]
      responses:
        '200':
          description: Lista de expedientes
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                    example: true
                  data:
                    type: array
                    items:
                      $ref: '#/components/schemas/Expediente'
        '401':
          description: No autorizado
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'

    post:
      tags:
        - Expedientes
      summary: Crear expediente militar
      security:
        - BearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateExpedienteRequest'
      responses:
        '201':
          description: Expediente creado exitosamente
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                    example: true
                  data:
                    $ref: '#/components/schemas/Expediente'
        '400':
          description: Error de validación
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '401':
          description: No autorizado
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /expedientes/{id}:
    get:
      tags:
        - Expedientes
      summary: Obtener expediente por ID
      security:
        - BearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Expediente encontrado
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                    example: true
                  data:
                    $ref: '#/components/schemas/Expediente'
        '404':
          description: Expediente no encontrado
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'

    put:
      tags:
        - Expedientes
      summary: Actualizar expediente
      security:
        - BearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateExpedienteRequest'
      responses:
        '200':
          description: Expediente actualizado exitosamente
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                    example: true
                  data:
                    $ref: '#/components/schemas/Expediente'
        '400':
          description: Error de validación
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '404':
          description: Expediente no encontrado
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'

    delete:
      tags:
        - Expedientes
      summary: Eliminar expediente
      security:
        - BearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Expediente eliminado exitosamente
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SuccessResponse'
        '404':
          description: Expediente no encontrado
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /expedientes/search:
    get:
      tags:
        - Expedientes
      summary: Buscar expedientes
      security:
        - BearerAuth: []
      parameters:
        - name: q
          in: query
          description: Término de búsqueda
          schema:
            type: string
        - name: grado
          in: query
          schema:
            type: string
            enum: [GRAL, CRL, TTE CRL, MY, CAP, TTE, STTE, TCO, SSOO, EC, TROPA]
        - name: cip
          in: query
          schema:
            type: string
      responses:
        '200':
          description: Resultados de búsqueda
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                    example: true
                  data:
                    type: array
                    items:
                      $ref: '#/components/schemas/Expediente'
...
  /permissions:
    get:
      tags:
        - Permissions
      summary: Listar todos los permisos disponibles
      description: Obtiene la lista completa de permisos disponibles en el sistema
      security:
        - BearerAuth: []
      responses:
        '200':
          description: Lista de permisos disponibles
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                    example: true
                  data:
                    type: object
                    properties:
                      expedientes:
                        type: array
                        items:
                          type: string
                        example: ["read", "create", "update", "delete"]
                      users:
                        type: array
                        items:
                          type: string
                        example: ["read", "create", "update", "delete"]
                      profiles:
                        type: array
                        items:
                          type: string
                        example: ["read", "create", "update", "delete"]
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
...
    LoginRequest:
      type: object
      required:
        - email
        - password
      properties:
        email:
          type: string
          format: email
          example: admin@sistema.mil
          description: Email del usuario en el sistema
        password:
          type: string
          minLength: 8
          example: admin123
          description: Contraseña del usuario
      example:
        email: admin@sistema.mil
        password: admin123
    LoginResponse:
      type: object
      properties:
        success:
          type: boolean
          example: true
        data:
          type: object
          properties:
            user:
              $ref: '#/components/schemas/User'
            access_token:
              type: string
              example: eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9.eyJzdWIiOiIxMjM0NTY3ODkwIiwibmFtZSI6IkpvaG4gRG9lIiwiaWF0IjoxNTE2MjM5MDIyfQ.SflKxwRJSMeKKF2QT4fwpMeJf36POk6yJV_adQssw5c
              description: Token JWT para autenticación (usar en header Authorization como 'Bearer {token}')
            refresh_token:
              type: string
              example: eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9...
              description: Token de renovación
            expires_at:
              type: string
              format: date-time
              example: 2025-11-04T20:00:00Z
              description: Fecha y hora de expiración del token
    ErrorResponse:
      type: object
      properties:
        success:
          type: boolean
          example: false
        error:
          type: string
          example: "Credenciales incorrectas"
...